# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ethio Pass API",
    "description": "Secure national e-KYC API platform for identity verification using biometrics, OCR, and liveness detection.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
    "/verify-id": {
      "post": {
        "summary": "Verify ID Document (OCR)",
        "description": "Upload an image of an Ethiopian National ID, passport, or driver's license to extract text using OCR.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EthiopianOCRVerificationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful OCR extraction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EthiopianOCRVerificationOutput"
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/liveness": {
      "post": {
        "summary": "Selfie Liveness Check",
        "description": "Analyze a selfie image to determine if it is of a live person or a spoofing attempt.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfieLivenessCheckInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful liveness check",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelfieLivenessCheckOutput"
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/match": {
      "post": {
        "summary": "Biometric Face Match",
        "description": "Compare two images (one from an ID, one from a selfie) to verify if they are the same person.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BiometricMatchInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful biometric match",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiometricMatchOutput"
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EthiopianOCRVerificationInput": {
        "type": "object",
        "properties": {
          "photoDataUri": {
            "type": "string",
            "format": "uri",
            "description": "A photo of an Ethiopian document as a data URI."
          }
        },
        "required": ["photoDataUri"]
      },
      "EthiopianOCRVerificationOutput": {
        "type": "object",
        "properties": {
          "extractedData": {
            "type": "string",
            "description": "The extracted data from the document."
          }
        }
      },
      "SelfieLivenessCheckInput": {
        "type": "object",
        "properties": {
          "photoDataUri": {
            "type": "string",
            "format": "uri",
            "description": "A selfie photo as a data URI."
          }
        },
        "required": ["photoDataUri"]
      },
      "SelfieLivenessCheckOutput": {
        "type": "object",
        "properties": {
          "isLive": {
            "type": "boolean",
            "description": "Whether the selfie is of a live person."
          },
          "confidence": {
            "type": "number",
            "description": "The confidence level of the liveness detection (0-1)."
          }
        }
      },
      "BiometricMatchInput": {
        "type": "object",
        "properties": {
          "idPhotoDataUri": {
            "type": "string",
            "format": "uri",
            "description": "A photo from an ID document as a data URI."
          },
          "selfiePhotoDataUri": {
            "type": "string",
            "format": "uri",
            "description": "A selfie photo as a data URI."
          }
        },
        "required": ["idPhotoDataUri", "selfiePhotoDataUri"]
      },
      "BiometricMatchOutput": {
        "type": "object",
        "properties": {
          "match": {
            "type": "boolean",
            "description": "Whether the two faces are a likely match."
          },
          "confidence": {
            "type": "number",
            "description": "A confidence score (0-1) for the match determination."
          },
          "reasoning": {
            "type": "string",
            "description": "A brief explanation for the decision."
          }
        }
      }
    }
  }
}